% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{activity_animal_csv_example}
\alias{activity_animal_csv_example}
\title{Dummy JAX Envision® animal-level activity CSV lines formatted the same as exported activity data.}
\format{
a character vector with 5 lines, the first line is a header.
}
\usage{
data(activity_animal_csv_example)
}
\description{
A dataset containing four lines of fabricated JAX Envision® activity data.
These lines follow the format of a JAX Envision® animal-level activity CSV.
When read properly, the lines for cage-level data have raw column titles as follows:
}
\details{
\itemize{
\item \code{start}. the date and time at the start of the aggregation bin (coerced from UTC to local time).
\item \code{start_date_local}. the start date (in the time zone in which the data were collected).
\item \code{start_time_local}. the start time (in the time zone in which the data were collected).
\item \code{study_code}. a unique code for each study.
\item \code{aggregation_seconds}. the number of seconds aggregated to generate this dataset (3600 is 1 hour).
\item \code{group_name}. a user-defined group name, often used to label experimental groups of interest.
\item \code{cage_name}. the name of the cage that the data represent.
\item \code{animals_cage_quantity}. the number of animals in the cage, sometimes called cage density or occupancy.
\item \code{light_cycle}. whether the data were collected in the light or dark cycle.
\item \code{animal.id}. the unique ID of the animal.
\item \code{strain}. the strain of animal.
\item \code{sex}. the sex of the animal.
\item \code{genotype}. the genotype of the animal.
\item \code{birth_date}. the date of birth for the animal.
\item \code{movement_animal_cm_s_hour}. cage-level movement in cm/s for a specified period of time (1 hour in this example)
\item \code{wheel_occupancy_animal_s.hour}. amount of time spent on the wheel at the cage level.
\item \code{food_occupancy_animal_s_hour}. amount of time spent in proximity to the food hopper at the cage level.
\item \code{water_occupancy_animal_s_hour}. amount of time spent in proximity to the water bottles at the cage level.
\item \code{tzone}. time zone of the dataset.
}
}
\keyword{datasets}
