% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{annotation_csv_example}
\alias{annotation_csv_example}
\title{Dummy JAX Envision® annotation CSV lines formatted the same as exported annotation data.}
\format{
a character vector with 5 lines, the first line is a header.
}
\usage{
data(annotation_csv_example)
}
\description{
A dataset containing four lines of dummy JAX Envision® annotation data.
These lines follow the format of a JAX Envision® annotation CSV.
When read properly, the lines for annotation data have column titles as follows:
}
\details{
\itemize{
\item \code{id}. the ID of the annotation, usually a number.
\item \code{created}. the date and time the annotation was created (coerced from UTC to local time).
\item \code{created_date_local}. the date the annotation was created (in the time zone in which the data were collected).
\item \code{created_time_local}. the time the annotation was created (in the time zone in which the data were collected).
\item \code{pin_start_date_local}. the date of the start of the period to which the annotation refers (in the time zone in which the data were collected).
\item \code{pin_start_time_local}. the time of the start of the period to which the annotation refers (in the time zone in which the data were collected).
\item \code{pin_end_date_local}. the date of the end of the period to which the annotation refers (in the time zone in which the data were collected).
\item \code{pin_end_time_local}. the time of the end of the period to which the annotation refers (in the time zone in which the data were collected).
\item \code{study_code}. a unique code for each study.
\item \code{group_name}. a user-defined group name, often used to label experimental groups of interest.
\item \code{cage_name}. the name of the cage that the data represent.
\item \code{creator}. the name of the user or bot who created the annotation.
\item \code{contents}. the annotation.
\item \code{reply_to}. if the annotation is a reply, the name of the user or bot to whom the reply is directed.
\item \code{hashtags}. any hashtags in the annotation.
\item \code{pin_start_time}. the date and time of the start of the period to which the annotation refers (in the time zone in which the data were collected).
\item \code{pin_end_time}. athe date and time of the end of the period to which the annotation refers (in the time zone in which the data were collected).
\item \code{tzone}. time zone of the dataset.
}
}
\keyword{datasets}
