% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacube_func.R
\name{build_datacube}
\alias{build_datacube}
\title{Build data cube}
\usage{
build_datacube(
  delta_array,
  fields,
  idu_geom,
  idu_field = "IDU_INDEX",
  as_stars = TRUE
)
}
\arguments{
\item{delta_array}{Data frame containing Envision delta array}

\item{fields}{Fields to include in the data cube}

\item{idu_geom}{IDU geometry saved as a sf object}

\item{idu_field}{Name containing IDU index}

\item{as_stars}{Return results as stars object}
}
\value{
datacube or stars object
}
\description{
Build data cube
}
