% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacube_func.R
\name{build_slice}
\alias{build_slice}
\title{Build slice}
\usage{
build_slice(
  delta_array,
  idu_geom,
  idu_field = "IDU_INDEX",
  fields,
  start_year = NULL,
  end_year = NULL
)
}
\arguments{
\item{end_year}{}
}
\value{
data slice
}
\description{
Builds component slice that are later combined into a single data cube
}
