% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{csv_column_defs}
\alias{csv_column_defs}
\title{Column definition}
\format{
a list with column type definitions for various types of Envision® csv exports.
}
\usage{
data(csv_column_defs)
}
\description{
Definitions for the columns of different Envision® csv types and versions.
}
\keyword{datasets}
