% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_func.R
\name{download_and_build_data_cube}
\alias{download_and_build_data_cube}
\title{Download and build data cube from delta array at given URL}
\usage{
download_and_build_data_cube(url, idu_shp = NULL, fields = NULL)
}
\arguments{
\item{url}{}
}
\description{
Download and build data cube from delta array at given URL
}
