% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_func.R
\name{download_delta_array}
\alias{download_delta_array}
\title{Download delta array from OSU BEE website}
\usage{
download_delta_array(
  url_slug = NULL,
  output_directory = NULL,
  url_base = "http://explorer.bee.oregonstate.edu/Topic/Deschutes/Outputs/Scenarios/",
  delete_temp_dir = FALSE,
  timeout = 1200
)
}
\arguments{
\item{timeout}{}
}
\value{
Temporary folder where delta array is saved
}
\description{
Download delta array from OSU BEE website
}
