% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envision_metadata.R
\name{envision_metadata}
\alias{envision_metadata}
\title{JAX Envision® Metadata}
\usage{
envision_metadata(
  study_name = "",
  tzone = NULL,
  lights_on = NULL,
  lights_off = NULL,
  study_url = NULL,
  override_tzassume = FALSE,
  override_timecheck = FALSE,
  force_lightcycle = FALSE,
  quietly = FALSE
)
}
\arguments{
\item{study_name}{The name of the study (used for default figure legends)}

\item{tzone}{The time zone of the study}

\item{lights_on}{the 24-hour time for lights-on as a \code{character},
\code{"HH:MM:SS"}}

\item{lights_off}{the 24-hour time for lights-off as a \code{character}}

\item{study_url}{the URL from the address bar of the study,
a \code{character}. NOTE: this can be any URL taken from within a study.}

\item{override_tzassume}{override the default behavior of assuming a time
zone? (default: \code{FALSE})}

\item{override_timecheck}{override the default behavior of assuming a time
zone? (default: \code{FALSE})}

\item{force_lightcycle}{force a light/dark cycle that is less than 8 hours
(defaults to \code{FALSE})}

\item{quietly}{suppress warning messages? (default: \code{FALSE})}
}
\description{
This function allows you to create a time-saving metadata object that can be
used as input in many of the \code{envisionR} functions. It checks the
inputted metadata to ensure they make sense.
}
\note{
In the United States, the most common time zones are: \code{US/Eastern},
\code{US/Central}, \code{US/Mountain}, and \code{US/Pacific}. Some more uncommon US time
zones are: \code{US/Alaska}, \code{US/Hawaii}, and \code{US/Arizona}.

For Europe, the time zone \code{GB} will work for Great Britain, the
time zone \code{WET} for Portugal, the time zone \code{CET} will work for much
of central Europe, and \code{EET} will work for Finland, the Baltics, and many
eastern European countries. For any other time zones, use the
\code{list_tzones()}

For any other time zones, the function \code{list_tzones()} will produce a
data frame containing all valid time zone designations in the \code{tz_name}
column.
}
