% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envisionR-package.R
\docType{package}
\name{envisionR-package}
\alias{envisionR}
\alias{envisionR-package}
\title{envisionR: JAX Envision® Data Wrangling and Analysis Package}
\description{
A package for importing, wrangling, analyzing, and visualizing data generated using the JAX Envision® platform.
}
\author{
\strong{Maintainer}: Michael Saul \email{michael.saul@jax.org} (\href{https://orcid.org/0000-0002-4916-8619}{ORCID})

}
\keyword{internal}
