% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_time_encoding.R
\name{experiment_time_encoding}
\alias{experiment_time_encoding}
\title{Experiment time encoding}
\usage{
experiment_time_encoding(
  rawtimes,
  reftimes,
  units = c("seconds", "minutes", "hours", "days", "weeks", "months", "years"),
  offset = 0
)
}
\arguments{
\item{rawtimes}{raw times to be changed to experiment times.}

\item{reftimes}{reference time or times for an experiment time encoding. Either length 1 or the same length as \code{rawtimes}.}

\item{units}{units for the experiment time encoding (hours, days, weeks, months, or years).}

\item{offset}{value given for the offset from 0 in the experiment in the same units as \code{units} (default: \code{0}).}
}
\value{
a \code{vector} of experiment time in the given units.
}
\description{
This function uses the date-time information in an annotation or an activity data set to compute experimental time.
Experimental time is useful in aligning disparate studies and for re-encoding by something like date of birth.
}
\keyword{Envision}
