% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_utc_offset.R
\name{get_utc_offset}
\alias{get_utc_offset}
\title{Get UTC Offset}
\usage{
get_utc_offset(ts, as_numeric = FALSE)
}
\arguments{
\item{ts}{Time stamp in local time}

\item{as_numeric}{Whether UTC offset should be returned as a \code{numeric} (default: \code{FALSE}).}
}
\value{
A \code{character} or \code{numeric} UTC offset.
}
\description{
This function computes UTC offset for a given time point.
Getting a UTC offset can be useful when daylight time starts or stops.
}
\examples{
get_utc_offset(ts = lubridate::ymd_hms("2024-06-01 12:00:00", tz = "US/Pacific"))
}
\keyword{Envision}
