% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_level_mean_sem.R
\name{group_level_mean_sem}
\alias{group_level_mean_sem}
\title{A Function to Compute Group-Level Summary Statistics on JAX Envision® activity CSVs}
\usage{
group_level_mean_sem(
  activity_data,
  var = "movement_mean_per_cage_cm_s_hour",
  by = "start",
  occupancy_normalize = TRUE,
  occupancy_weighted = TRUE,
  na.rm = TRUE,
  quietly = FALSE
)
}
\arguments{
\item{activity_data}{a cage-level activity dataset.}

\item{var}{the activity metric to summarize.
(defaults to \code{"movement_mean_per_cage_cm_s_hour"})}

\item{by}{time varible to summarize by (defaults to \code{"start"})}

\item{occupancy_normalize}{should the function normalize by cage occupancy?
(defaults to \code{TRUE})}

\item{occupancy_weighted}{should the function weight by cage occupancy?
(defaults to \code{TRUE})}

\item{na.rm}{should the function remove \code{NA} values?
(default: \code{TRUE})}

\item{quietly}{should the function suppress warnings?
(default: \code{FALSE})}
}
\value{
summary statistics for the dataset
}
\description{
This function computes mean, standard error, and error bars for activity data.
}
\examples{
# Getting dummy dataset
data("activity_cage_data_example")

# Computing summary statistics
x <- group_level_mean_sem(activity_cage_data_example)

# Looking at the summary statistics
head(x)
}
\keyword{Envision}
