% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightdark_times.R
\name{lightdark_times}
\alias{lightdark_times}
\title{Get light/dark times}
\usage{
lightdark_times(
  lights_on_time,
  lights_off_time,
  start_date,
  end_date,
  tzone = "UTC"
)
}
\arguments{
\item{lights_on_time}{Time of lights-on, a string formatted as \code{"HH:MM"}}

\item{lights_off_time}{Time of lights-off, a string formatted as \code{"HH:MM"}}

\item{start_date}{First dates in the series of dates, a string formatted as \code{"YYYY-MM-DD"}}

\item{end_date}{Last date in the series of dates, a string formatted as \code{"YYYY-MM-DD"}}

\item{tzone}{The time zone, which should match a time zone given in the \code{list_tzones()} function.}
}
\value{
A \code{tibble} with lights-off and lights-on times for all dates, useful for constructing figures.
}
\description{
This function returns a \code{tibble} of lights-off and lights-on times.
}
\examples{
lightdark_times(
  lights_on_time = "06:00",
  lights_off_time = "18:00",
  start_date = "2024-06-02",
  end_date = "2024-06-03",
  tzone = "US/Pacific"
)
}
\keyword{Envision}
