% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightdark_violin_plot.R
\name{lightdark_violin_plot}
\alias{lightdark_violin_plot}
\title{A \code{ggplot2} template for a faceted violin plot with activity in light and dark}
\usage{
lightdark_violin_plot(
  activity_data,
  metadata = NULL,
  visualize_on,
  yvar,
  colors = NULL,
  y_axis_label = NULL,
  quietly = FALSE
)
}
\arguments{
\item{activity_data}{activity data frame.}

\item{metadata}{a metadata object as made by \code{envision_metadata()}.}

\item{visualize_on}{the variable to visualize on for colors,
a \code{character} of length 1 (defaults to \code{group_name}).
NOTE: as of 2024-08-30, the ability to visualize on grouping variables
other than \code{group_name} or \code{minoccupancy} has not been implemented.}

\item{yvar}{the variable for the y-axis,
a \code{character} of length 1.}

\item{colors}{user-defined colors for the plot.}

\item{y_axis_label}{user-specified y-axis label.}

\item{quietly}{should the function suppress warnings? (default: \code{FALSE})}
}
\value{
a \code{ggplot2} based violin plot with inset box-and-whiskers.
}
\description{
This function returns a template \code{ggplot2} object that is a violin plot with
activity in the light and dark.
}
\examples{
# Making dummy metadata
metadata <- envision_metadata(
  study_name = "A",
  tzone = "US/Pacific",
  lights_on = "06:00:00",
  lights_off = "18:00:00",
  study_url = "https://envision.jax.org/org/1001/study/1002/overview"
)

# Getting dummy dataset
data("activity_cage_data_example")

# Plotting dummy data
lightdark_violin_plot(
  activity_data = activity_cage_data_example,
  metadata = metadata,
  visualize_on = "group_name",
  yvar = "movement_mean_per_cage_cm_s_hour"
)
}
\keyword{Envision}
