% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_tzones.R
\name{list_tzones}
\alias{list_tzones}
\title{A Function to list time zones compatible with the JAX Envision® software}
\usage{
list_tzones()
}
\value{
A \code{data.frame} with compatible time zones in the \code{tz_name} column.
}
\description{
This function allows you list possible time zones for the JAX Envision® software.
}
\examples{
list_tzones()
}
\keyword{Envision}
