% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ddin_font.R
\name{load_ddin_font}
\alias{load_ddin_font}
\title{A helper function to load the D-DIN font}
\usage{
load_ddin_font(quietly = FALSE)
}
\arguments{
\item{quietly}{should the function suppress warnings? (default: \code{FALSE})}
}
\value{
No return value.
}
\description{
This function loads an open source version of DIN 1451. This font was
originally prepared for the company Datto and licensed under the OFL v1.1.
The \href{https://github.com/amcchord/datto-d-din}{D-DIN font} is available to
download and redistribute freely. This function does not return any value.
}
\keyword{Envision}
