% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_envision_url.R
\name{make_envision_url}
\alias{make_envision_url}
\title{Make a URL for the JAX Envision® platform}
\usage{
make_envision_url(
  org,
  study,
  cage,
  vidstart,
  windowstart_h = 12,
  windowend_h = 12,
  metricstab = "cage",
  videostream = "overlay"
)
}
\arguments{
\item{org}{Organization number of the video.}

\item{study}{Study number of the video.}

\item{cage}{Cage number for the video.}

\item{vidstart}{Start time of the video (a \code{POSIXct} object).}

\item{windowstart_h}{Hours before the start of the video for the window.}

\item{windowend_h}{Hours after the start of the video for the window.}

\item{metricstab}{Which metrics tab to return (either \code{"cage"}, \code{"animal"}, or \code{"alpha"}).}

\item{videostream}{Which video stream to return (note: for now, \code{"overlay"} is the default).}
}
\value{
A \code{character} vector containing the JAX Envision® URL.
}
\description{
This function creates a URL for the JAX Envision® platform.
URLs are useful for highlighting specific moments of video identified by an analysis.
}
\examples{
make_envision_url(1001, 1001, 1001, as.POSIXct("2024-06-12 07:45:00", tz = "US/Central"))
}
\keyword{Envision}
