% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_func.R
\name{palette_list}
\alias{palette_list}
\title{Pull palette data from XML}
\usage{
palette_list(xml_file = NULL)
}
\arguments{
\item{xml_file}{}
}
\value{
list of palettes
}
\description{
Pull palette data from XML
}
