% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_func.R
\name{plot_raster}
\alias{plot_raster}
\title{Plot raster}
\usage{
plot_raster(
  raster,
  pal_lookup = NULL,
  pal_table = NULL,
  extent = NULL,
  focus = NULL,
  options = c("Roads", "Zones", "Legend"),
  maxcell = 1e+05
)
}
\arguments{
\item{options}{character vector of map elements}

\item{maxcell}{}
}
\description{
Plot raster
}
\details{
options include Roads, Zones, and Legend
}
