% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacube_func.R
\name{pull_yr0_slice}
\alias{pull_yr0_slice}
\title{Pull yr0 slice}
\usage{
pull_yr0_slice(idu_geom, fields, idu_field, start_year, end_year)
}
\arguments{
\item{end_year}{}
}
\value{
IDU slice at year 0
}
\description{
Used to pull intial conditions from idu reference layer
}
