% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_annotation_csv.R
\name{read_annotation_csv}
\alias{read_annotation_csv}
\title{A Function to Read JAX Envision® annotation CSVs}
\usage{
read_annotation_csv(csv, tz = NULL)
}
\arguments{
\item{csv}{File path for the Envision® annotation CSV.}

\item{tz}{Time zone for the study. If left as \code{NULL}, the function will attempt to assume a time zone and throw a warning.}
}
\value{
A \code{tibble} with annotation data optimally formatted for downstream analysis.
}
\description{
This function allows you to read annotation CSVs generated by the JAX Envision® software.
}
\examples{
# Writing test CSV file
annotation_csv <- tempfile("testannotation", fileext = ".csv")
readr::write_lines(annotation_csv_example, file = annotation_csv)

# Reading in test CSV file
annotation <- read_annotation_csv(csv = annotation_csv, tz = "US/Pacific")
dplyr::glimpse(annotation)
}
\keyword{Envision}
