% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_envision_csv.R
\name{read_envision_csv}
\alias{read_envision_csv}
\title{A Function to Read JAX Envision® CSVs}
\usage{
read_envision_csv(csv, tz = NULL)
}
\arguments{
\item{csv}{CSV for your JAX Envision® file.}
}
\description{
This function allows you to read CSVs generated by the JAX Envision® software.
}
\examples{
read_envision_csv()
}
\keyword{Envision}
