% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reltime.R
\name{reltime}
\alias{reltime}
\title{Relative time encoding}
\usage{
reltime(
  rawtimes,
  reftimes,
  units = c("seconds", "minutes", "hours", "days", "weeks", "months", "years"),
  offset = 0
)
}
\arguments{
\item{rawtimes}{raw times to be changed to relative times.}

\item{reftimes}{reference time or times for a relative time encoding. Either length 1 or the same length as \code{rawtimes}.}

\item{units}{units for the relative time encoding (hours, days, weeks, months, or years).}

\item{offset}{value given for the offset from 0 in the relative in the same units as \code{units} (default: \code{0}).}
}
\value{
a \code{vector} of relative time in the given units.
}
\description{
This function uses the date-time information in an annotation or an activity data set to compute relative time.
Relative time is useful in aligning disparate studies and for re-encoding by something like date of birth.
}
\keyword{Envision}
