% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaghetti_plot.R
\name{spaghetti_plot}
\alias{spaghetti_plot}
\title{A \code{ggplot2} template for a spaghetti plot with light-dark times}
\usage{
spaghetti_plot(
  activity_data,
  metadata = NULL,
  lights_on = NULL,
  lights_off = NULL,
  visualize_on = "group_name",
  xvar = "start",
  yvar = NULL,
  colors = NULL,
  lightdark = TRUE,
  darkphase_col = "#CCCCCC",
  darkphase_alpha = 0.5,
  facet = TRUE,
  occupancy_norm = TRUE,
  x_axis_label = NULL,
  y_axis_label = NULL,
  xlims = NULL,
  ylims = NULL,
  quietly = FALSE
)
}
\arguments{
\item{activity_data}{activity data frame.}

\item{metadata}{a metadata object as made by \code{envision_metadata()}.}

\item{lights_on}{a lights-on time (formatted as \code{"HH:MM"}).}

\item{lights_off}{a lights-off time (formatted as \code{"HH:MM"}).}

\item{visualize_on}{the variable to visualize on for colors,
a \code{character} of length 1 (defaults to \code{group_name}).
NOTE: as of 2024-08-30, the ability to visualize on grouping variables
other than \code{group_name} has not been implemented.}

\item{xvar}{the variable for the x-axis,
a \code{character} of length 1 (defaults to \code{start})}

\item{yvar}{the variable for the y-axis,
a \code{character} of length 1.}

\item{colors}{the colors for the plot.}

\item{lightdark}{should the plot return light-dark cycle indicators on the
plot background? (default: \code{TRUE})}

\item{darkphase_col}{the color for the dark phase background (default:
\code{"#CCCCCC"})}

\item{darkphase_alpha}{the transparency alpha for the dark phase background.
Transparency allows grid lines to be visible in the background. (default:
\code{0.5})}

\item{facet}{should the plot facet on the grouping variables? (default:
\code{TRUE})}

\item{occupancy_norm}{should the plot include data normalized for cage
occupancy? (default: \code{TRUE})}

\item{x_axis_label}{user-specified x-axis label.}

\item{y_axis_label}{user-specified y-axis label.}

\item{xlims}{user-specified x-axis limits.}

\item{ylims}{user-specified y-axis limits.}

\item{quietly}{should the function suppress warnings? (default: \code{FALSE})}
}
\value{
a \code{ggplot2} based spaghetti plot.
}
\description{
This function returns a \code{ggplot2} object that is a spaghetti plot with
light-dark times grayed out.
}
\examples{
# Making dummy metadata
metadata <- envision_metadata(
  study_name = "A",
  tzone = "US/Pacific",
  lights_on = "06:00:00",
  lights_off = "18:00:00",
  study_url = "https://envision.jax.org/org/1001/study/1002/overview"
)

# Getting dummy dataset
data("activity_cage_data_example")

# Plotting dummy data
spaghetti_plot(
  activity_data = activity_cage_data_example,
  metadata = metadata,
  yvar = "movement_mean_per_cage_cm_s_hour",
  facet = TRUE
)
}
\keyword{Envision}
