% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_jax_pub.R
\name{theme_jax_pub}
\alias{theme_jax_pub}
\title{A \code{ggplot2} theme for JAX branded content}
\usage{
theme_jax_pub(
  base_size = 11,
  base_family = "D-DIN Expanded",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{base_size}{base font size in pts.}

\item{base_family}{base font family}

\item{base_line_size}{base size for line elements}

\item{base_rect_size}{base size for rect elements}
}
\value{
a \code{ggplot2} theme for automated publishable plots for JAX.
}
\description{
This function returns a \code{ggplot2} theme with the D-DIN font, a JAX
branded font.
}
\keyword{Envision}
