% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_func.R
\name{update_raster}
\alias{update_raster}
\title{Update IDU raster}
\usage{
update_raster(
  raster,
  datacube,
  field = 1,
  year = 1,
  extent = NULL,
  queryA = "",
  queryB = ""
)
}
\arguments{
\item{query}{}
}
\description{
Update IDU raster
}
