% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocity_unit_convert.R
\name{velocity_unit_convert}
\alias{velocity_unit_convert}
\title{Convenience function for simple velocity unit conversion}
\usage{
velocity_unit_convert(x, units_in = "cm/s", units_out = "cm/s")
}
\arguments{
\item{x}{The velocity as a \code{numeric} vector}

\item{units_in}{The velocity units in}

\item{units_out}{The velocity units out

The units should be in \code{distance/time} format.
Valid values for distance are:
\itemize{
\item \code{mm}: millimeters
\item \code{cm}: centimeters
\item \code{m}: meters
\item \code{km}: kilometers
}
Valid values for time are:
\itemize{
\item \code{s}: second
\item \code{m}: minute
\item \code{h}: hour
}}
}
\description{
This function allows you to convert between velocity units.
}
