% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weeks_postnatal_seq.R
\name{weeks_postnatal_seq}
\alias{weeks_postnatal_seq}
\title{Make a sequence of weeks postnatal}
\usage{
weeks_postnatal_seq(dob, n_weeks)
}
\arguments{
\item{dob}{Date of birth of the animal in a \code{ymd} format.}

\item{n_weeks}{Number of weeks as a \code{numeric}.}
}
\value{
A named vector of dates where the names are the postnatal weeks.
}
\description{
This function computes a vector of dates that start each postnatal weeks after a birth date.
This is often useful for axis legends.
}
\examples{
weeks_postnatal_seq()
}
\keyword{Envision}
