% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_sem.R
\name{weighted_sem}
\alias{weighted_sem}
\title{A Function to Compute Weighted Standard Error of the Mean JAX Envision® activity CSVs}
\usage{
weighted_sem(x, w, na.rm = FALSE)
}
\arguments{
\item{x}{vector of type \code{numeric} for computing weighted standard error}

\item{w}{weights to apply to \code{x}}

\item{na.rm}{Should the function remove \code{NA} obervations?
(defaults to \code{FALSE})}
}
\value{
the weighted standard error of the dataset.
}
\description{
This function computes standard error with weights.
}
\examples{
# Making an example dataset
x <- c(0, 0, 3, -3, 0, 0, 0, -0.75, -0.75, 0.75, 0.75)
wts <- c(3, 3, 2, 2, 2, 1, 1, 1, 1, 1, 1)

# Computing weighted standard error (should return 0.5)
weighted_sem(x, wts)

# Computing weighted standard error with NAs added to the dataset
# The function should return 0.5
x_na <- c(x, NA)
wts_na <- c(wts, 1)
weighted_sem(x_na, wts_na, na.rm = TRUE)

}
\keyword{Envision}
