% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_activity_csv.R
\docType{data}
\name{csv_column_defs}
\alias{csv_column_defs}
\title{Column definition}
\format{
a list with both cage-level and animal-level column specifications
}
\usage{
data(csv_column_defs)
}
\description{
Definitions for the columns of different Envision® csv types and versions.
}
\keyword{datasets}
