% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_utc_offset.R
\name{get_utc_offset}
\alias{get_utc_offset}
\title{Get UTC Offset}
\usage{
get_utc_offset(ts, ts_utc = NULL, as_numeric = FALSE)
}
\arguments{
\item{ts}{time stamp in local time}

\item{ts_utc}{time stamp in UTC time (default: \code{NULL})}

\item{as_numeric}{whether UTC offset should be returned as a \code{numeric} (default: \code{FALSE}).}
}
\value{
A \code{character} or \code{numeric} UTC offset.
}
\description{
This function computes UTC offset for a given time point.
Getting a UTC offset can be useful when daylight time starts or stops.
}
\examples{
# Use with no second time stamp and an explicit time zone
get_utc_offset(ts = as.POSIXct("2024-06-01 12:00:00", tzone = "US/Pacific"))

# Use with a second time stamp and no explicit time zone.
get_utc_offset(ts = as.POSIXct("2024-06-01 12:00:00"),
               ts_utc = as.POSIXct("2024-06-01 05:00:00"))
}
\keyword{Envision}
