% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_envision_url.R
\name{make_envision_url}
\alias{make_envision_url}
\title{Make a URL for the JAX Envision速 platform}
\usage{
make_envision_url(
  org,
  study,
  cage,
  vidstart,
  windowstart_h = 12,
  windowend_h = 12,
  metricstab = "cage",
  videostream = "overlay",
  url_base = "https://app.murine.net/org/"
)
}
\arguments{
\item{org}{Organization number of the video.}

\item{study}{Study number of the video.}

\item{cage}{Cage number for the video.}

\item{vidstart}{Start time of the video (usually a \code{timepoint} or \code{instant} created with the \code{lubridate} package).}

\item{metricstab}{Which metrics tab to return (either \code{"cage"}, \code{"animal"}, or \code{"alpha"}).}

\item{videostream}{Which video stream to return (note: for now, \code{"overlay"} is the default).}

\item{url_base}{What URL base is used for the Envision速 URL: NOTE: the default of \code{"https://app.murine.net/org/"} is from the original version of the app.}

\item{widnowstart_h}{Hours after the start of the video for the window.}
}
\value{
A \code{character} vector containing the JAX Envision速 URL.
}
\description{
This function creates a URL for the JAX Envision速 platform.
URLs are useful for highlighting specific moments of video identified by an analysis.
}
\examples{
make_envision_url()
}
\keyword{Envision}
