% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_activity_csv.R
\name{read_activity_csv}
\alias{read_activity_csv}
\title{A Function to Read JAX Envision® activity CSVs}
\usage{
read_activity_csv(csv, tz = NULL, metrics = NULL)
}
\arguments{
\item{csv}{File path for the Envision® CSV.}

\item{tz}{Time zone for the study. If left as \code{NULL}, the function will attempt to assume a time zone and throw a warning.}

\item{metrics}{Should the function expect cage (\code{"cage"}) or animal
(\code{"animal"}) level activity metrics? Default: \code{NULL} tries to
identify whether the csv is animal or cage level.}
}
\value{
A \code{tibble} with experimental data optimally formatted for downstream analysis.
}
\description{
This function allows you to read activity CSVs generated by the JAX Envision® software.
}
\examples{
# Writing test CSV file for an example
activity_csv <- tempfile("testactivity", fileext = ".csv")
readr::write_lines(activity_csv_example, file = activity_csv)

# Reading in test activity CSV file
activity <- read_activity_csv(csv = activity_csv, tz = "US/Pacific")

# Glimpsing test activity CSV file
dplyr::glimpse(activity)

# Removing the activity file
file.remove(activity_csv)
}
\keyword{Envision}
