% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_activity_csv.R
\name{read_activity_csv}
\alias{read_activity_csv}
\title{A Function to Read JAX Envision® activity CSVs}
\usage{
read_activity_csv(
  csv,
  tz = NULL,
  occupancy_normalize = FALSE,
  metrics = c("cage", "animal")
)
}
\arguments{
\item{csv}{File path for the Envision® CSV.}

\item{tz}{Time zone for the study. If left as \code{NULL}, the function will attempt to assume a time zone and throw a warning.}

\item{occupancy_normalize}{Should the function normalize activity by cage occupancy? Default: \code{FALSE}.}
}
\value{
A \code{tibble} with experimental data optimally formatted for downstream analysis.
}
\description{
This function allows you to read activity CSVs generated by the JAX Envision® software.
}
\examples{
read_envision_csv()
}
\keyword{Envision}
